requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-restframework==0.0.1
djangorestframework==3.12.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==1.5
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
